adapt vtol attitude control class to new vtol type classes
