+++ src/modules/vtol_att_control/vtol_att_control_params.c
@@ -142,3 +142,11 @@ PARAM_DEFINE_FLOAT(VT_PROP_EFF,0.0f);
  */
 PARAM_DEFINE_FLOAT(VT_ARSP_LP_GAIN,0.3f);
 
+/**
+ * VTOL Type (Tailsitter=0, Tiltrotor=1)
+ *
+ * @min 0
+ * @max 1
+ * @group VTOL Attitude Control
+ */
+PARAM_DEFINE_INT32(VT_TYPE, 0);
